requirements.txt
# Core Web Framework
fastapi==0.124.4
uvicorn[standard]==0.40.0

# Data Validation & Settings
pydantic==2.12.5
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.45
alembic==1.18.1
asyncpg==0.31.0

# HTTP Client
httpx==0.28.1

# YAML Config
pyyaml==6.0.2

# JWT Authentication - SECURITY FIX: Upgraded from 3.3.0 due to CVE-2024-33663, CVE-2024-33664, CVE-2025-61152
# Alternative: Consider migrating to PyJWT which is more actively maintained
python-jose[cryptography]==3.5.0

# Caching
cachetools>=5.5.0

# Structured Logging
structlog>=24.0.0

# Testing
pytest==8.4.0
pytest-asyncio==1.3.0
pytest-cov==4.1.0
